# A tiny optimization to ImLineClosestPoint. Removed unnecessary sqrtf call.