This might be useful if some widget makes a lot of calls to *ImLineClosestPoint*. The call to *sqrtf* was eliminated while keeping the behavior of the functions precisely the same. 

According to the google benchmark, it is roughly x2.4 faster now.

    Benchmark                             Time           CPU Iterations
    -------------------------------------------------------------------
    ImLineClosestPoint_original        4867 ns       5000 ns     100000
    ImLineClosestPoint_optimized       1981 ns       1950 ns     344615