@ocornut 2018-02-22
Thank you @podgorskiy ! Could you rename `ab_lenSqr` to `ab_len_sqr` and rebase-squash the commit it creates into one?