@stfx 2018-02-22
@ocornut btw Github allows you to do all that yourself too (edit PR using the online editor unless opted out by author and "squash and merge" in the merge dropdown button) which might be more efficient for small changes like these