@ocornut 2018-02-22
@stfx Interesting, I didn't know there was a way to edit online (didn't see it here!). Thanks!
I agree those changes are relatively easy to do! Sometimes I do them, sometimes I request it as a way to sensitize people to making PR with less issues, which would ease the job for me.