@podgorskiy 2018-02-22
Fixed.
@stfx, I've moved that line.